+++ Tp-webII-2daParte/templates_c/f25cb1010f336917e14baa7409de3f124c4de30a_0.file.escuderiaList.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.2.1, created on 2023-10-12 23:41:05
+/* Smarty version 4.2.1, created on 2023-10-13 01:42:09
   from 'C:\xampp\htdocs\Tp-webII-2daParte\templates\escuderiaList.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_652867f10a1801_84589671',
+  'unifunc' => 'content_6528845109f0f0_87408844',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     'f25cb1010f336917e14baa7409de3f124c4de30a' => 
     array (
       0 => 'C:\\xampp\\htdocs\\Tp-webII-2daParte\\templates\\escuderiaList.tpl',
-      1 => 1697146862,
+      1 => 1697154126,
       2 => 'file',
     ),
   ),
@@ -23,11 +23,13 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:footer.tpl' => 1,
   ),
 ),false)) {
-function content_652867f10a1801_84589671 (Smarty_Internal_Template $_smarty_tpl) {
+function content_6528845109f0f0_87408844 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_checkPlugins(array(0=>array('file'=>'C:\\xampp\\htdocs\\Tp-webII-2daParte\\libs\\smarty-4.2.1\\libs\\plugins\\modifier.truncate.php','function'=>'smarty_modifier_truncate',),));
 $_smarty_tpl->_subTemplateRender("file:header.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
-$_smarty_tpl->_subTemplateRender("file:form_altaEscuderia.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
-?>
+if ((isset($_SESSION['USER_ID']))) {?>
+    <?php $_smarty_tpl->_subTemplateRender("file:form_altaEscuderia.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
+}?>
+
 
 <table class="table table-striped">
     <thead>
@@ -66,11 +68,12 @@ $_smarty_tpl->tpl_vars['escuderias']->do_else = false;
 </td>
                 <td><?php echo smarty_modifier_truncate($_smarty_tpl->tpl_vars['escuderias']->value->pos_equipos,35);?>
 </td>
+            <?php if ((isset($_SESSION['USER_ID']))) {?>
                 <td><a href='editequipoform/<?php echo $_smarty_tpl->tpl_vars['escuderias']->value->id;?>
 ' type='button' class='btn btn-info'>Editar</a></td>
                 <td><a href='eliminar/<?php echo $_smarty_tpl->tpl_vars['escuderias']->value->id;?>
 ' type='button' class='btn btn-danger'>Borrar</a></td>
-                
+            <?php }?>
             </span> 
         </td>
